Add support for internal iteration to avoid heap allocations
This adds alternative query implementations that rely on internal instead of
external iteration to avoid the necessary heap allocations for reifying
the stack within the selection iterators.

While this does increase the probability of stack overflow, it does significantly
reduce the allocator bottle neck when multiple threads perform queries in parallel.