Upgrade heapless to v0.7 and fix resulting compile error
Heapless v0.7 was released ~10 months ago with support for the Const
Generics MVP, allowing for constant-length arrays to be used in generics.

This commit updates the heapless dependency to 0.7.

The only thing that seems to have broken when I made this change was the
SmallHeap type used in the nearest_neighbor algorithm, so I changed the
signature for the heapless::binary_heap::BinaryHeap type accordingly.

With v0.7, heapless no longer has a heapless::consts::U32 array length,
nor the ArrayLength generic.  Instead, heapless::binary_heap::BinaryHeap
now makes use of a const usize generic type parameter, which is the last
parameter.

To support heapless v0.7, I only had to move the Max parameter and make
use of the const generic length parameter.  Since N = U32 (which means
"32, as an unsigned integer") in heapless v0.6 can now be represented as
N = 32_usize in Rust 1.51+, this should be equivalent.  The tests passed,
at least. :)

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>