Fix AABB::from_points which relied on implementation details of AABB::new_empty (#171)
- [x] I agree to follow the project's [code of
conduct](https://github.com/georust/geo/blob/master/CODE_OF_CONDUCT.md).
- [x] I added an entry to `rstar/CHANGELOG.md` if knowledge of this
change could be valuable to users.
---

This method does not just need an empty AABB, but actually requires that
the lower/upper corners use max/min values of the point coordinates.

This change therefore open-codes this to keep these invariants and the
code that relies on them close together.

A changelog entry should not be necessary as we luckily did not yet
release the regression.

Closes #170