Revert back to min/max representation of empty AABB (#184)
- [x] I agree to follow the project's [code of
conduct](https://github.com/georust/geo/blob/master/CODE_OF_CONDUCT.md).
- [x] I added an entry to `rstar/CHANGELOG.md` if knowledge of this
change could be valuable to users.
---

This reverts back to the min/max representation of empty AABB due to
regressions introduced by the numerically more tame but not invariant
for merging representation.

To avoid regressing #161, it also adds a single check to avoid computing
the distance to an empty envelope (of the root node of an empty tree).
Integer coordinates are always prone to overflow but in the empty case
we are forcing this onto the caller whereas every non-empty tree will
have AABB that the caller supplied and where we can reasonably ask them
to control for overflow or use a custom `Point` impl based on saturating
arithmetic.

Fixes #183